toggle between gas and oil production. refactor tables into components, types, interface, and utils
