+++ src/types/interface.ts
@@ -1,0 +1,8 @@
+
+export interface Projection {
+  oilFieldName: string;
+  year: number;
+  productionOil: number | null;
+  productionGas: number | null;
+  emission: number | null;
+}
