+++ src/types/types.ts
@@ -1,0 +1,6 @@
+
+export type OilFieldDataset = Record<
+  string, Record<
+    string, { productionOil?: number; productionGas?: number; emission?: number; }
+  >
+>;
