+++ src/utils/projections.ts
@@ -1,0 +1,75 @@
+import { Projection } from "../types/interface";
+import { OilFieldDataset } from "../types/types";
+
+export function isStillProducing(
+  yearlyData: Record<
+    string, { productionOil?: number; productionGas?: number; emission?: number; }
+  >,
+  resourceKey: "productionOil" | "productionGas"): boolean {
+  const years = Object.keys(yearlyData)
+    .map(Number)
+    .sort((a, b) => b - a);
+  const latestYear = years[0];
+  return yearlyData[latestYear]?.[resourceKey] !== undefined;
+}
+
+export function calculateAverage(
+  yearlyData: Record<
+    string, { productionOil?: number; productionGas?: number; emission?: number; }
+  >,
+  resourceKey: "productionOil" | "productionGas"): number | null {
+  const years = Object.keys(yearlyData)
+    .map(Number)
+    .filter((year) => yearlyData[year]?.[resourceKey] !== undefined)
+    .sort((a, b) => b - a)
+    .slice(0, 5);
+
+  const values = years.map((year) => yearlyData[year]![resourceKey]!);
+
+  if (values.length === 0) return null;
+
+  const avg = values.reduce((a, b) => a + b, 0) / values.length;
+  return avg;
+}
+
+export function productionProjections(data: OilFieldDataset): Projection[] {
+  const projections: Projection[] = [];
+  const projectionStart = 2023;
+  const projectionEnd = 2040;
+  const annualDeclineRate = 0.1;
+
+  for (const [fieldName, yearlyData] of Object.entries(data)) {
+    const oilActive = isStillProducing(yearlyData, "productionOil");
+    const gasActive = isStillProducing(yearlyData, "productionGas");
+
+    let projectedOil = oilActive
+      ? calculateAverage(yearlyData, "productionOil")
+      : null;
+    let projectedGas = gasActive
+      ? calculateAverage(yearlyData, "productionGas")
+      : null;
+
+    for (let year = projectionStart; year <= projectionEnd; year++) {
+      projections.push({
+        oilFieldName: fieldName,
+        year,
+        productionOil: projectedOil !== null ? parseFloat(projectedOil.toFixed(2)) : null,
+        productionGas: projectedGas !== null ? parseFloat(projectedGas.toFixed(2)) : null,
+        emission: null,
+      });
+
+      if (projectedOil !== null) {
+        projectedOil *= 1 - annualDeclineRate;
+        if (projectedOil < 0.01) projectedOil = 0;
+      }
+
+      if (projectedGas !== null) {
+        projectedGas *= 1 - annualDeclineRate;
+        if (projectedGas < 0.01) projectedGas = 0;
+      }
+    }
+  }
+
+  return projections;
+}
+
